Correction de qql bugs : Penser a supprimer autre close
